Mid 0.0.1 -> 0.0.2 Scene Camera Feature In progress
Tons of Additions and Bug Fixes
Added Memory Management

+++ Wingine (Standalone)/Attributes.cs
@@ -100,7 +100,7 @@ namespace Wingine
         readonly double min;
         readonly double max;
 
-        public Range(double min = double.MinValue, double max = double.MaxValue)
+        public Range(double min = int.MinValue, double max = int.MaxValue)
         {
             this.min = min;
             this.max = max;
